Add test cases for bad TOML files
