+++ tests/messages/toml-test-cases/bad.extractor.toml
@@ -1,0 +1,2 @@
+[extractors]
+custom = { module = "mypackage.module", func = "myfunc" }
