+++ tests/messages/toml-test-cases/bad.extractors-not-a-dict.toml
@@ -1,0 +1,1 @@
+[[extractors]]
