+++ tests/messages/toml-test-cases/bad.just-a-mapping.toml
@@ -1,0 +1,3 @@
+[mapping]
+method = "jinja2"
+pattern = "**.html"
