+++ tests/messages/toml-test-cases/bad.mapping-not-a-dict.toml
@@ -1,0 +1,1 @@
+mappings = [8]
