+++ tests/messages/toml-test-cases/bad.mappings-not-a-list.toml
@@ -1,0 +1,1 @@
+mappings = "python"
