+++ tests/messages/toml-test-cases/bad.missing-extraction-method.toml
@@ -1,0 +1,2 @@
+[[mappings]]
+pattern = ["xyz/**.py", "foo/**.py"]
