+++ tests/messages/toml-test-cases/bad.multiple-mappings-not-a-list.toml
@@ -1,0 +1,10 @@
+[mappings]
+method = "genshi"
+pattern = "**/templates/**.html"
+include_attrs = ""
+
+[mappings]
+method = "genshi"
+pattern = "**/templates/**.txt"
+template_class = "genshi.template:TextTemplate"
+encoding = "latin-1"
