+++ tests/messages/toml-test-cases/bad.non-string-extraction-method.toml
@@ -1,0 +1,2 @@
+[[mappings]]
+method = 42
