+++ tests/messages/toml-test-cases/bad.pattern-type-2.toml
@@ -1,0 +1,3 @@
+[[mappings]]
+method = "big snake"
+pattern = [42]
