+++ tests/messages/toml-test-cases/bad.pattern-type.toml
@@ -1,0 +1,3 @@
+[[mappings]]
+method = "big snake"
+pattern = 2048
