+++ tests/messages/toml-test-cases/bad.standalone-with-babel-prefix.toml
@@ -1,0 +1,2 @@
+[babel.extractors]
+custom = "mypackage.module:myfunc"
